requirements.txt
altgraph==0.16.1
certifi==2018.8.24
chardet==3.0.4
docopt==0.6.2
ffmpeg-python==0.1.16
future==0.16.0
idna==2.7
macholib==1.11
numpy==1.15.1
opencv-python==3.4.3.18
pefile==2018.8.8
PyInstaller==3.4
pywin32-ctypes==0.2.0
requests==2.19.1
urllib3==1.23
